TEMPLATE_CORE/USER/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "user-core"
version = "0.1.0"
description = "Abstract SQLAlchemy 2.x base models for unified user management systems"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["user_core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
